Change dicom series managment and image from dicom conversion
Some OT volumes have the same Series Instance UID, so the conversion from dicom to image fails
So now, volumes with the same Series Instance UID are set in separate DicomSerie entries in the table (3D Dicom CT slices are not volume, so they are not concerned)

Moreover, conversion from dicom to image for OT volumes failed because ITK looked for the images with the same Series Instance UID in the folder.
I decided to separate DicomSerie in different folder to avoid this problem. So now, when we insert a dicom serie entry, the id correspond to the folder where the files are store

To avoid code repetition, I prefer to use gatetools to convert dicom to image. I use >=v0.8.7 to be able to manage 3D Dicom volume with 1 slice